Moved parsing for coloring to different files for ease of management
